logging package - morgan,compression, cors chore: change dev start command modify login response fix registeration response - toJSON - remove password from response fix toDo delete method

+++ src/services/todoService.ts
@@ -1,7 +1,5 @@
 import { ITodo, Todo } from "../models/todoModel";
 
-// TODO: add try catch block
-
 export const getTodoDetails = async (todoId: string, userId: string) => {
   const todoDetails = await Todo.findOne({ id: todoId, userId });
   if (!todoDetails) throw new Error("Todo does not exists");
@@ -9,7 +7,6 @@ export const getTodoDetails = async (todoId: string, userId: string) => {
   return todoDetails;
 };
 
-// TODO: Add Pagination
 export const getAllTodos = async (userId: string) => {
   const todoDetails = await Todo.find({ userId });
   if (!todoDetails) throw new Error("Todo does not exists For this user");
@@ -21,14 +18,22 @@ export const createTodo = async (userId: string, todo: ITodo) => {
   return await Todo.create({ ...todo, userId });
 };
 
-export const updateTodo = async (todoId: string, todo: ITodo) => {
+export const updateTodo = async (
+  todoId: string,
+  userId: string,
+  todo: ITodo
+) => {
+  try {
+    return await Todo.updateOne({ _id: todoId, userId }, todo);
+  } catch (error) {
+    console.debug(error);
+  }
   return await Todo.updateOne({ _id: todoId }, todo);
 };
 
-// TODO : need to check if toDo belongs to user - same for all
-export const deleteTodo = async (todoId: string) => {
+export const deleteTodo = async (userId: string, todoId: string) => {
   try {
-    await Todo.deleteOne({ _id: todoId });
+    await Todo.deleteOne({ _id: todoId, userId });
   } catch (error) {
     console.debug(error);
   }
